# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaseFacilityType =
  (typeof CaseFacilityType)[keyof typeof CaseFacilityType];

export const CaseFacilityType = {
  term_loan: "term_loan",
  working_capital: "working_capital",
  letter_of_credit: "letter_of_credit",
  bank_guarantee: "bank_guarantee",
  overdraft: "overdraft",
} as const;

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  draft: "draft",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Case {
  id: number;
  borrowerName: string;
  cin?: string;
  pan?: string;
  facilityType: CaseFacilityType;
  facilityAmount: number;
  sector: string;
  rmName: string;
  status: CaseStatus;
  /** 0-100 percent progress */
  memoProgress: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateCaseBodyFacilityType =
  (typeof CreateCaseBodyFacilityType)[keyof typeof CreateCaseBodyFacilityType];

export const CreateCaseBodyFacilityType = {
  term_loan: "term_loan",
  working_capital: "working_capital",
  letter_of_credit: "letter_of_credit",
  bank_guarantee: "bank_guarantee",
  overdraft: "overdraft",
} as const;

export interface CreateCaseBody {
  borrowerName: string;
  cin?: string;
  pan?: string;
  facilityType: CreateCaseBodyFacilityType;
  facilityAmount: number;
  sector: string;
  rmName: string;
}

export type UpdateCaseBodyFacilityType =
  (typeof UpdateCaseBodyFacilityType)[keyof typeof UpdateCaseBodyFacilityType];

export const UpdateCaseBodyFacilityType = {
  term_loan: "term_loan",
  working_capital: "working_capital",
  letter_of_credit: "letter_of_credit",
  bank_guarantee: "bank_guarantee",
  overdraft: "overdraft",
} as const;

export type UpdateCaseBodyStatus =
  (typeof UpdateCaseBodyStatus)[keyof typeof UpdateCaseBodyStatus];

export const UpdateCaseBodyStatus = {
  draft: "draft",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateCaseBody {
  borrowerName?: string;
  cin?: string;
  pan?: string;
  facilityType?: UpdateCaseBodyFacilityType;
  facilityAmount?: number;
  sector?: string;
  rmName?: string;
  status?: UpdateCaseBodyStatus;
}

export type MemoSectionConfidence =
  (typeof MemoSectionConfidence)[keyof typeof MemoSectionConfidence];

export const MemoSectionConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
  pending: "pending",
} as const;

export interface MemoSection {
  id: number;
  caseId: number;
  sectionKey: string;
  sectionTitle: string;
  content: string;
  confidence: MemoSectionConfidence;
  isReviewed: boolean;
  isLocked: boolean;
  updatedAt: string;
}

export interface UpdateSectionBody {
  content?: string;
  isReviewed?: boolean;
  isLocked?: boolean;
}

export interface GenerateResponse {
  message: string;
  caseId: number;
}

export type RiskFlagSeverity =
  (typeof RiskFlagSeverity)[keyof typeof RiskFlagSeverity];

export const RiskFlagSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface RiskFlag {
  id: number;
  caseId: number;
  riskType: string;
  severity: RiskFlagSeverity;
  description: string;
  mitigation?: string;
  isAcknowledged: boolean;
}

export interface DashboardStats {
  totalCases: number;
  draftsInProgress: number;
  pendingReview: number;
  approvedThisMonth: number;
  avgDraftTimeHours: number;
  timeSavedHours: number;
}

export interface ActivityItem {
  id: number;
  caseId: number;
  borrowerName: string;
  action: string;
  actor: string;
  timestamp: string;
}

export interface StatusBreakdownItem {
  status: string;
  count: number;
  label: string;
}

export type ListCasesParams = {
  status?: ListCasesStatus;
  search?: string;
};

export type ListCasesStatus =
  (typeof ListCasesStatus)[keyof typeof ListCasesStatus];

export const ListCasesStatus = {
  draft: "draft",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
} as const;