# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all credit memo cases
 */
export const ListCasesQueryParams = zod.object({
  status: zod.enum(["draft", "in_review", "approved", "rejected"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListCasesResponseItem = zod.object({
  id: zod.number(),
  borrowerName: zod.string(),
  cin: zod.string().optional(),
  pan: zod.string().optional(),
  facilityType: zod.enum([
    "term_loan",
    "working_capital",
    "letter_of_credit",
    "bank_guarantee",
    "overdraft",
  ]),
  facilityAmount: zod.number(),
  sector: zod.string(),
  rmName: zod.string(),
  status: zod.enum(["draft", "in_review", "approved", "rejected"]),
  memoProgress: zod.number().describe("0-100 percent progress"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListCasesResponse = zod.array(ListCasesResponseItem);

/**
 * @summary Create a new credit memo case
 */
export const CreateCaseBody = zod.object({
  borrowerName: zod.string(),
  cin: zod.string().optional(),
  pan: zod.string().optional(),
  facilityType: zod.enum([
    "term_loan",
    "working_capital",
    "letter_of_credit",
    "bank_guarantee",
    "overdraft",
  ]),
  facilityAmount: zod.number(),
  sector: zod.string(),
  rmName: zod.string(),
});

/**
 * @summary Get a credit memo case by ID
 */
export const GetCaseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCaseResponse = zod.object({
  id: zod.number(),
  borrowerName: zod.string(),
  cin: zod.string().optional(),
  pan: zod.string().optional(),
  facilityType: zod.enum([
    "term_loan",
    "working_capital",
    "letter_of_credit",
    "bank_guarantee",
    "overdraft",
  ]),
  facilityAmount: zod.number(),
  sector: zod.string(),
  rmName: zod.string(),
  status: zod.enum(["draft", "in_review", "approved", "rejected"]),
  memoProgress: zod.number().describe("0-100 percent progress"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a credit memo case
 */
export const UpdateCaseParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCaseBody = zod.object({
  borrowerName: zod.string().optional(),
  cin: zod.string().optional(),
  pan: zod.string().optional(),
  facilityType: zod
    .enum([
      "term_loan",
      "working_capital",
      "letter_of_credit",
      "bank_guarantee",
      "overdraft",
    ])
    .optional(),
  facilityAmount: zod.number().optional(),
  sector: zod.string().optional(),
  rmName: zod.string().optional(),
  status: zod.enum(["draft", "in_review", "approved", "rejected"]).optional(),
});

export const UpdateCaseResponse = zod.object({
  id: zod.number(),
  borrowerName: zod.string(),
  cin: zod.string().optional(),
  pan: zod.string().optional(),
  facilityType: zod.enum([
    "term_loan",
    "working_capital",
    "letter_of_credit",
    "bank_guarantee",
    "overdraft",
  ]),
  facilityAmount: zod.number(),
  sector: zod.string(),
  rmName: zod.string(),
  status: zod.enum(["draft", "in_review", "approved", "rejected"]),
  memoProgress: zod.number().describe("0-100 percent progress"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a credit memo case
 */
export const DeleteCaseParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get all memo sections for a case
 */
export const ListSectionsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListSectionsResponseItem = zod.object({
  id: zod.number(),
  caseId: zod.number(),
  sectionKey: zod.string(),
  sectionTitle: zod.string(),
  content: zod.string(),
  confidence: zod.enum(["high", "medium", "low", "pending"]),
  isReviewed: zod.boolean(),
  isLocked: zod.boolean(),
  updatedAt: zod.coerce.date(),
});
export const ListSectionsResponse = zod.array(ListSectionsResponseItem);

/**
 * @summary Update (edit) a memo section
 */
export const UpdateSectionParams = zod.object({
  id: zod.coerce.number(),
  sectionKey: zod.coerce.string(),
});

export const UpdateSectionBody = zod.object({
  content: zod.string().optional(),
  isReviewed: zod.boolean().optional(),
  isLocked: zod.boolean().optional(),
});

export const UpdateSectionResponse = zod.object({
  id: zod.number(),
  caseId: zod.number(),
  sectionKey: zod.string(),
  sectionTitle: zod.string(),
  content: zod.string(),
  confidence: zod.enum(["high", "medium", "low", "pending"]),
  isReviewed: zod.boolean(),
  isLocked: zod.boolean(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Trigger AI generation for a memo case
 */
export const GenerateMemoParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateMemoResponse = zod.object({
  message: zod.string(),
  caseId: zod.number(),
});

/**
 * @summary Get risk flags for a case
 */
export const ListRiskFlagsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListRiskFlagsResponseItem = zod.object({
  id: zod.number(),
  caseId: zod.number(),
  riskType: zod.string(),
  severity: zod.enum(["high", "medium", "low"]),
  description: zod.string(),
  mitigation: zod.string().optional(),
  isAcknowledged: zod.boolean(),
});
export const ListRiskFlagsResponse = zod.array(ListRiskFlagsResponseItem);

/**
 * @summary Get overall dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalCases: zod.number(),
  draftsInProgress: zod.number(),
  pendingReview: zod.number(),
  approvedThisMonth: zod.number(),
  avgDraftTimeHours: zod.number(),
  timeSavedHours: zod.number(),
});

/**
 * @summary Get recent case activity timeline
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  caseId: zod.number(),
  borrowerName: zod.string(),
  action: zod.string(),
  actor: zod.string(),
  timestamp: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);

/**
 * @summary Get case counts broken down by status
 */
export const GetStatusBreakdownResponseItem = zod.object({
  status: zod.string(),
  count: zod.number(),
  label: zod.string(),
});
export const GetStatusBreakdownResponse = zod.array(
  GetStatusBreakdownResponseItem,
);