# lib/api-zod/src/generated/types/createCaseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCaseBodyFacilityType } from "./createCaseBodyFacilityType";

export interface CreateCaseBody {
  borrowerName: string;
  cin?: string;
  pan?: string;
  facilityType: CreateCaseBodyFacilityType;
  facilityAmount: number;
  sector: string;
  rmName: string;
}