# lib/api-zod/src/generated/types/generateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateResponse {
  message: string;
  caseId: number;
}