# lib/api-zod/src/generated/types/memoSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoSectionConfidence } from "./memoSectionConfidence";

export interface MemoSection {
  id: number;
  caseId: number;
  sectionKey: string;
  sectionTitle: string;
  content: string;
  confidence: MemoSectionConfidence;
  isReviewed: boolean;
  isLocked: boolean;
  updatedAt: Date;
}