# lib/api-zod/src/generated/types/riskFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFlagSeverity } from "./riskFlagSeverity";

export interface RiskFlag {
  id: number;
  caseId: number;
  riskType: string;
  severity: RiskFlagSeverity;
  description: string;
  mitigation?: string;
  isAcknowledged: boolean;
}