# lib/api-zod/src/generated/types/updateCaseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCaseBodyFacilityType } from "./updateCaseBodyFacilityType";
import type { UpdateCaseBodyStatus } from "./updateCaseBodyStatus";

export interface UpdateCaseBody {
  borrowerName?: string;
  cin?: string;
  pan?: string;
  facilityType?: UpdateCaseBodyFacilityType;
  facilityAmount?: number;
  sector?: string;
  rmName?: string;
  status?: UpdateCaseBodyStatus;
}