# lib/api-zod/src/generated/types/updateSectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreditGuard AI — LLM-Powered Credit Memo Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSectionBody {
  content?: string;
  isReviewed?: boolean;
  isLocked?: boolean;
}